04-prj-color-palette-generator/ColorPaletteFlaskApp/pyproject.toml
[tool.poetry]
name = "colorpaletteflaskapp"
version = "0.1.0"
description = ""
authors = ["Noah Kim <dongwoo_333@hanmail.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
openai = "^0.27.6"
python-dotenv = "^1.0.0"
autopep8 = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
